+++ src/hooks/useInfiniteScroll.js
@@ -1,0 +1,47 @@
+import { useEffect, useState, useRef } from "react";
+
+export default function useInfiniteScroll(datas) {
+  const observerRef = useRef(null);
+  const [items, setItems] = useState([]);
+  const [loading, setLoading] = useState(true);
+
+  useEffect(() => {
+    getData();
+    observer.observe(observerRef.current);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [datas]);
+
+  useEffect(() => {
+    if (items.length === 0) return;
+    if (items.length === datas.length) {
+      observer.disconnect();
+      alert("더 불러올 데이터가 없습니다");
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [items]);
+
+  const getData = async () => {
+    try {
+      setItems((prev) => [
+        ...prev,
+        ...datas.slice(prev.length, prev.length + 6),
+      ]);
+    } catch (error) {
+      alert(error.message);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  const handleIntersection = (entries) => {
+    if (entries[0].isIntersecting) {
+      getData();
+    }
+  };
+
+  const observer = new IntersectionObserver(handleIntersection, {
+    threshold: 0.5,
+  });
+
+  return { items, loading, observerRef };
+}
